+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:example/valuebar_position.dart';
 import 'package:flutter/material.dart';
 import 'package:geekyants_flutter_gauges/geekyants_flutter_gauges.dart';
 
@@ -6,56 +5,80 @@ void main() {
   runApp(
     const MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: MyValueBarPosition(),
+      home: RadialGaugeExample(),
     ),
   );
 }
 
-class MyGaugeExample extends StatefulWidget {
-  const MyGaugeExample({Key? key}) : super(key: key);
+///
+/// The following code  is a Simple Example of [LinearGauge] Widget.
+/// You can customize the [LinearGauge] Widget as per your need.
+///
+class LinearGaugeExample extends StatefulWidget {
+  const LinearGaugeExample({Key? key}) : super(key: key);
 
   @override
-  State<MyGaugeExample> createState() => _MyGaugeExampleState();
+  State<LinearGaugeExample> createState() => _LinearGaugeExampleState();
 }
 
-class _MyGaugeExampleState extends State<MyGaugeExample> {
-  double value = 50;
-  bool inverseRulers = false;
-
+class _LinearGaugeExampleState extends State<LinearGaugeExample> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            ElevatedButton(
-                onPressed: () {
-                  inverseRulers = !inverseRulers;
-                  setState(() {});
-                },
-                child: const Text("Inverse Rulers")),
-            Container(
-              color: Colors.pink.withOpacity(0.1),
-              child: LinearGauge(
-                gaugeOrientation: GaugeOrientation.horizontal,
-                valueBar: [
-                  ValueBar(
-                    value: value,
-                    offset: 20,
-                    position: ValueBarPosition.top,
-                  )
-                ],
-                enableGaugeAnimation: true,
-                rulers: RulerStyle(
-                  rulerPosition: RulerPosition.bottom,
-                  inverseRulers: inverseRulers,
-                  showLabel: inverseRulers,
-                ),
-              ),
+        child: LinearGauge(
+          gaugeOrientation: GaugeOrientation.horizontal,
+          enableGaugeAnimation: true,
+          rulers: RulerStyle(
+            rulerPosition: RulerPosition.bottom,
+          ),
+          pointers: const [
+            Pointer(
+              value: 50,
+              shape: PointerShape.circle,
             ),
           ],
         ),
+      ),
+    );
+  }
+}
+
+///
+/// The following code  is a Simple Example of [RadialGauge] Widget.
+/// You can customize the [RadialGauge] Widget as per your need.
+///
+
+class RadialGaugeExample extends StatefulWidget {
+  const RadialGaugeExample({super.key});
+
+  @override
+  State<RadialGaugeExample> createState() => _RadialGaugeExampleState();
+}
+
+class _RadialGaugeExampleState extends State<RadialGaugeExample> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      backgroundColor: Colors.white,
+      body: RadialGauge(
+        track: RadialTrack(
+            color: Colors.grey,
+            start: 0,
+            end: 100,
+            trackStyle: TrackStyle(
+                showLastLabel: false,
+                secondaryRulerColor: Colors.grey,
+                secondaryRulerPerInterval: 3)),
+        needlePointer: [
+          NeedlePointer(
+            value: 30,
+            color: Colors.red,
+            tailColor: Colors.black,
+            tailRadius: 0,
+            needleStyle: NeedleStyle.flatNeedle,
+          ),
+        ],
       ),
     );
   }