+++ lib/src/radial_gauge/pointer/needle_pointer_painter.dart
@@ -1,4 +1,5 @@
 import 'dart:math';
+import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import '../../../geekyants_flutter_gauges.dart';
 
@@ -8,10 +9,12 @@ class RenderNeedlePointer extends RenderBox {
     required double value,
     required Color color,
     required RadialGauge radialGauge,
+    required ValueChanged<double>? onChanged,
     required double needleHeight,
     required double tailRadius,
     required double needleWidth,
     required NeedleStyle needleStyle,
+    required bool isInteractive,
     required Color tailColor,
     required LinearGradient gradient,
 
@@ -20,10 +23,12 @@ class RenderNeedlePointer extends RenderBox {
         _color = color,
         _radialGauge = radialGauge,
         _gradient = gradient,
+        _onChanged = onChanged,
         _needleHeight = needleHeight,
         _tailRadius = tailRadius,
         _tailColor = tailColor,
         _needleStyle = needleStyle,
+        _isInteractive = isInteractive,
         _needleWidth = needleWidth,
         super();
 
@@ -33,7 +38,21 @@ class RenderNeedlePointer extends RenderBox {
     if (_gradient == gradient) return;
     _gradient = gradient;
     markNeedsPaint();
+    markNeedsLayout();
   }
+
+  ValueChanged<double>? get onChanged => _onChanged;
+  ValueChanged<double>? _onChanged;
+  set onChanged(ValueChanged<double>? value) {
+    if (value == _onChanged) {
+      return;
+    }
+    _onChanged = value;
+    markNeedsPaint();
+    markNeedsLayout();
+  }
+
+  //getter and setter for value
 
   RadialGauge get getRadialGauge => _radialGauge!;
   RadialGauge? _radialGauge;
@@ -41,7 +60,23 @@ class RenderNeedlePointer extends RenderBox {
     if (_radialGauge == radialGauge) return;
     _radialGauge = radialGauge;
     markNeedsPaint();
+    markNeedsLayout();
   }
+
+  // Sets the Interaction for [RenderNeedlePointer].
+  set setIsInteractive(bool value) {
+    if (value == _isInteractive) {
+      return;
+    }
+
+    _isInteractive = value;
+    markNeedsPaint();
+    markNeedsLayout();
+  }
+
+  // Gets the Interaction assigned to [RenderNeedlePointer].
+  bool get isInteractive => _isInteractive;
+  bool _isInteractive;
 
   NeedleStyle get getNeedleStyle => _needleStyle;
   NeedleStyle _needleStyle;
@@ -49,6 +84,7 @@ class RenderNeedlePointer extends RenderBox {
     if (_needleStyle == needleStyle) return;
     _needleStyle = needleStyle;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   double get getValue => _value;
@@ -57,6 +93,7 @@ class RenderNeedlePointer extends RenderBox {
     if (_value == value) return;
     _value = value;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   double get getNeedleWidth => _needleWidth;
@@ -65,6 +102,7 @@ class RenderNeedlePointer extends RenderBox {
     if (_needleWidth == needleWidth) return;
     _needleWidth = needleWidth;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   double get getTailRadius => _tailRadius;
@@ -73,6 +111,7 @@ class RenderNeedlePointer extends RenderBox {
     if (_tailRadius == tailRadius) return;
     _tailRadius = tailRadius;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   double get getNeedleHeight => _needleHeight;
@@ -81,6 +120,7 @@ class RenderNeedlePointer extends RenderBox {
     if (_needleHeight == needleHeight) return;
     _needleHeight = needleHeight;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   Color get getTailColor => _tailColor;
@@ -89,6 +129,7 @@ class RenderNeedlePointer extends RenderBox {
     if (_tailColor == tailColor) return;
     _tailColor = tailColor;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   Color get getColor => _color;
@@ -97,27 +138,34 @@ class RenderNeedlePointer extends RenderBox {
     if (_color == color) return;
     _color = color;
     markNeedsPaint();
+    markNeedsLayout();
   }
 
   @override
-  bool hitTest(BoxHitTestResult result, {required Offset position}) {
-    return false;
-  }
-
-  @override
-  Size computeDryLayout(BoxConstraints constraints) {
-    return constraints.constrain(constraints.biggest);
+  bool hitTestSelf(Offset position) {
+    Offset calulatedPosition = localToGlobal(position);
+    if (needlePointerRect.contains(calulatedPosition)) {
+      return true;
+    } else {
+      return false;
+    }
+    // Return false if the position is outside the needlePointerRect
   }
 
   @override
   void performLayout() {
-    size = computeDryLayout(constraints);
+    // size = Size(_needleWidth + _tailRadius, _needleHeight);
+    size = Size(constraints.maxWidth, constraints.maxHeight);
   }
+
+  late Path needlePointerRect;
 
   @override
   void paint(PaintingContext context, Offset offset) {
     final canvas = context.canvas;
-    final center = offset;
+    offset = Offset(size.width * getRadialGauge.xCenterCoordinate + offset.dx,
+        size.height * getRadialGauge.yCenterCoordinate + offset.dy);
+    final center = Offset(offset.dx + _needleWidth, offset.dy + _needleHeight);
 
     Rect circle = Rect.fromLTWH(offset.dx - getTailRadius / 2,
         offset.dy - getTailRadius / 2, _tailRadius, _tailRadius);
@@ -134,16 +182,19 @@ class RenderNeedlePointer extends RenderBox {
     double startAngle = (_radialGauge!.track.startAngle - 180) * (pi / 180);
     double endAngle = (_radialGauge!.track.endAngle - 180) * (pi / 180);
 
-    final maxH = size.shortestSide / 2 * getRadialGauge.radiusFactor;
+    final maxH = size.shortestSide / 2 * getRadialGauge.radiusFactor -
+        2 * getRadialGauge.track.thickness;
 
-    final double needleMultiplier = _needleHeight.clamp(0, maxH - 40);
+    final double needleMultiplier = _needleHeight.clamp(0, maxH);
 
     final double angle = startAngle + (value / 100) * (endAngle - startAngle);
 
-    double needleEndX = center.dx + needleMultiplier * cos(angle);
-    double needleEndY = center.dy + needleMultiplier * sin(angle);
-    double needleStartX = center.dx;
-    double needleStartY = center.dy;
+    double needleEndX =
+        center.dx + needleMultiplier * cos(angle) - _needleWidth;
+    double needleEndY =
+        center.dy + needleMultiplier * sin(angle) - _needleHeight;
+    double needleStartX = center.dx - _needleWidth;
+    double needleStartY = center.dy - _needleHeight;
 
     final needlePaint = Paint()
       ..color = _color
@@ -163,17 +214,19 @@ class RenderNeedlePointer extends RenderBox {
     needlePath.lineTo(needleStartX + getTailRadius / 2 * cos(angle - pi / 2),
         needleStartY - (getTailRadius / 2) * sin(angle + pi / 2));
 
-    double temp1 = 30;
-    needlePath.moveTo(offset.dx + temp1 * cos(angle + pi / 2),
-        offset.dy - temp1 * sin(angle - pi / 2));
-    needlePath.lineTo(needleStartX + temp1 * cos(angle - pi / 2),
-        needleStartY - (temp1) * sin(angle + pi / 2));
+    double needleWidth = _needleWidth;
+    needlePath.moveTo(offset.dx + needleWidth * cos(angle + pi / 2),
+        offset.dy - needleWidth * sin(angle - pi / 2));
+    needlePath.lineTo(needleStartX + needleWidth * cos(angle - pi / 2),
+        needleStartY - (needleWidth) * sin(angle + pi / 2));
 
     needlePath.lineTo(needleEndX, needleEndY);
     // Offset c = Offset(offset.dx, offset.dy - getTailRadius);
     // canvas.drawCircle(c, 100, Paint()..color = Colors.black);
     needlePath.close();
 
+    needlePointerRect = needlePath;
+    // canvas.drawPath(needlePath, Paint()..color = Colors.green);
 // Needle  Pointer paint
     if (getNeedleStyle == NeedleStyle.gaugeNeedle) {
       canvas.drawPath(needlePath, needlePaint);
@@ -182,6 +235,7 @@ class RenderNeedlePointer extends RenderBox {
       //  Simple Needle
       canvas.drawLine(Offset(needleStartX, needleStartY),
           Offset(needleEndX, needleEndY), needlePaint);
+      canvas.drawPath(circlePath, Paint()..color = _tailColor);
     }
   }
 
@@ -191,3 +245,26 @@ class RenderNeedlePointer extends RenderBox {
     return newValue;
   }
 }
+
+// @override
+// void applyPaintTransform(RenderBox child, Matrix4 transform) {
+//   if (child is RenderNeedlePointer) {
+//     final centerX = size.width / 2;
+//     final centerY = size.height / 2;
+
+//     transform.translate(centerX, centerY);
+//     double value = calculateValueAngle(
+//         child.getValue, getRadialGauge.track.start, getRadialGauge.track.end);
+//     double startAngle = (getRadialGauge.track.startAngle - 180) * (pi / 180);
+//     double endAngle = (getRadialGauge.track.endAngle - 180) * (pi / 180);
+//     double angle = startAngle + (value / 100) * (endAngle - startAngle);
+//     double toRotateAngle = angle - (pi / 2);
+//     transform.rotateZ(toRotateAngle); // Specify the rotation in radians
+//     transform.translate(
+//         -centerX - child.getNeedleWidth / 2 - child.getTailRadius / 2,
+//         -centerY - child.getNeedleHeight + child.getTailRadius / 2);
+
+//     super.applyPaintTransform(child, transform);
+//     markNeedsLayout();
+//   }
+// }